Move MSSQL helpers to MSSQL project
