+++ Source/EventFlow.MsSql.Tests/Helpers/MsSqlDatabase.cs
@@ -24,7 +24,7 @@
 using System;
 using System.Data.SqlClient;
 
-namespace EventFlow.TestHelpers.MsSql
+namespace EventFlow.MsSql.Tests.Helpers
 {
     public class MsSqlDatabase : IMsSqlDatabase
     {
