+++ Source/EventFlow.MsSql.Tests/IntegrationTests/SnapshotStores/MsSqlSnapshotStoreTests.cs
@@ -26,8 +26,8 @@ using System.Threading;
 using EventFlow.Extensions;
 using EventFlow.MsSql.Extensions;
 using EventFlow.MsSql.SnapshotStores;
+using EventFlow.MsSql.Tests.Helpers;
 using EventFlow.TestHelpers;
-using EventFlow.TestHelpers.MsSql;
 using EventFlow.TestHelpers.Suites;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
